# todo-backend/app/Models/TodoList.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\HasMany;

/**
 * @OA\Schema()
 */
class TodoList extends Model
{
    use HasFactory;
    protected $fillable = ['name'];

    public function todos(): HasMany
    {
        return $this->hasMany(Todo::class);
    }

    /**
     * @OA\Property(format="int32")
     *
     * @var int
     */
    public $id;

    /**
     * @OA\Property(format="string")
     *
     * @var string
     */
    public $name;
}